Module Assumptions

val traverse : Global.indirect_accessor -> Names.Label.t -> Constr.constr -> Names.GlobRef.Set_env.t * Names.GlobRef.Set_env.t option Names.GlobRef.Map_env.t * (Names.Label.t * Constr.rel_context * Constr.types) list Names.GlobRef.Map_env.t

Collects all the objects on which a term directly relies, bypassing kernel opacity, together with the recursive dependence DAG of objects.

WARNING: some terms may not make sense in the environment, because they are sealed inside opaque modules. Do not try to do anything fancy with those terms apart from printing them, otherwise demons may fly out of your nose.

NOTE: this function is used in the plugin paramcoq.

val assumptions : ?add_opaque:bool -> ?add_transparent:bool -> Global.indirect_accessor -> TransparentState.t -> Names.GlobRef.t -> Constr.constr -> Constr.types Printer.ContextObjectMap.t

Collects all the assumptions (optionally including opaque definitions) on which a term relies (together with their type). The above warning of traverse also applies.