Module CSet.Hashcons

Create hash-consing for sets. The hashing function provided must be compatible with the comparison function.

Parameters

module M : OrderedType

module _ : Hashcons.HashedType with type t = M.t

Signature

type t = Stdlib.Set.Make(M).t

Type of objects to hashcons.

type table

Type of hashconsing tables

val generate : unit -> table

This create a hashtable of the hashconsed objects.

val hcons : table -> t Hashcons.f

Perform the hashconsing of the given object within the table, and returns the hash.

val stats : table -> Hashset.statistics

Recover statistics of the hashconsing table.