Module Class_tactics.Search

val eauto_tac : (Hints.Modes.t * TransparentState.t) -> ?unique:bool -> only_classes:bool -> best_effort:bool -> ?strategy:search_strategy -> depth:Int.t option -> dep:bool -> Hints.hint_db list -> unit Proofview.tactic

Note: there might be stuck goals due to mode declarations remaining even in case of success of the tactic.