Module Control

Global control of Rocq.

exception Timeout

Used to convert signals to exceptions

val enable_thread_delay : bool Stdlib.ref

Will periodically call Thread.delay if set to true

val interrupt : bool Stdlib.ref

Rocq interruption: set the following boolean reference to interrupt Rocq (it eventually raises Break, simulating a Ctrl-C)

val check_for_interrupt : unit -> unit

Use this function as a potential yield function. If interrupt has been set, il will raise Sys.Break.

val timeout : float -> ('a -> 'b) -> 'a -> ('b, Exninfo.info) Stdlib.result

timeout n f x tries to compute Ok (f x), and if it fails to do so before n seconds, returns Error info instead (where info contains the backtrace of the timeout exception).

type kilowords = {

kilowords : Stdlib.Int64.t;

}

val alloc_limit : kilowords -> ('a -> 'b) -> 'a -> ('b * kilowords, Exninfo.info) Stdlib.result

type timeout = {

timeout : 'a 'b. float -> ('a -> 'b) -> 'a -> ('b, Exninfo.info) Stdlib.result;

}

Set a particular timeout function; warning, this is an internal API and it is scheduled to go away.

val set_timeout : timeout -> unit

val protect_sigalrm : ('a -> 'b) -> 'a -> 'b

protect_sigalrm f x computes f x, but if SIGALRM is received during that computation, the signal handler is executed only once the computation is terminated. Otherwise said, it makes the execution of f atomic w.r.t. handling of SIGALRM.

This is useful for example to prevent the implementation of `Timeout` to interrupt I/O routines, generating ill-formed output.