Module Coqcargs

type compilation_mode =

| BuildVo

| BuildVos

| BuildVok

Compilation modes:

BuildVo : process statements and proofs (standard compilation), and also output an empty .vos file and .vok file

BuildVos : process statements, and discard proofs, and load .vos files for required libraries

BuildVok : like BuildVo, but load .vos files for required libraries

When loading the .vos version of a required library, if the file exists but is empty, then we attempt to load the .vo version of that library. This trick is useful to avoid the need for the user to compile .vos version when an up to date .vo version is already available.

type glob_output =

| NoGlob

| GlobFile of string

type t = {

compilation_mode : compilation_mode;

compile_file : string option;

compilation_output_name : string option;

beautify : bool;

glob_out : glob_output option;

output_context : bool;

}

val default : t

val parse : string list -> t