Module Coqloop

The Rocq toplevel loop.

A buffer for the character read from a channel. We store the command * entered to be able to report errors without pretty-printing.

type input_buffer

The input buffer of stdin.

val top_buffer : input_buffer

val set_prompt : (unit -> string) -> unit

val coqloop_feed : Feedback.feedback -> unit

Toplevel feedback printer.

val ml_toplevel_state : Vernac.State.t option Stdlib.ref

State tracked while in the OCaml toplevel

val ml_toplevel_include_ran : bool Stdlib.ref

Whether the "include" file was already run at least once

val loop : state:Vernac.State.t -> Vernac.State.t

The main loop

val run : opts:Coqargs.t -> state:Vernac.State.t -> unit

Main entry point of Rocq: read and execute vernac commands.