Module Declare.Info

Information for a declaration, interactive or not, includes parameters shared by mutual constants

type t

val make : ?poly:PolyFlags.t -> ?kind:Decls.logical_kind -> ?udecl:UState.universe_decl -> ?scope:Locality.definition_scope -> ?clearbody:bool -> ?hook:Hook.t -> ?typing_flags:Declarations.typing_flags -> ?user_warns:Globnames.extended_global_reference UserWarn.with_qf -> ?ntns:Metasyntax.notation_interpretation_decl list -> unit -> t

Note that opaque doesn't appear here as it is not known at the start of the proof in the interactive case.