Module Environ

Unsafe environments. We define here a datatype for environments. Since typing is not yet defined, it is not possible to check the informations added in environments, and that is why we speak here of ``unsafe'' environments.

Environments have the following components:

a context for de Bruijn variables

a context for de Bruijn variables vm values

a context for section variables and goal assumptions

a context for section variables and goal assumptions vm values

a context for global constants and axioms

a context for inductive definitions

a set of universe constraints

a flag telling if Set is, can be, or cannot be set impredicative

type link_info =

| Linked of string

| NotLinked

Linking information for the native compiler

type key = int CEphemeron.key option Stdlib.ref

type named_context_val

type rel_context_val

type env

Type of global environments.

type rewrule_not_allowed =

| Symb

| Rule

exception RewriteRulesNotAllowed of rewrule_not_allowed

val oracle : env -> Conv_oracle.oracle

val set_oracle : env -> Conv_oracle.oracle -> env

val eq_named_context_val : named_context_val -> named_context_val -> bool

val empty_env : env

val universes : env -> UGraph.t

val rel_context : env -> Constr.rel_context

val rel_context_val : env -> rel_context_val

val named_context : env -> Constr.named_context

val named_context_val : env -> named_context_val

val set_universes : UGraph.t -> env -> env

val set_qualities : QGraph.t -> env -> env

val qualities : env -> QGraph.t

val typing_flags : env -> Declarations.typing_flags

val is_impredicative_set : env -> bool

val type_in_type : env -> bool

val ignore_elim_constraints : env -> bool

val deactivated_guard : env -> bool

val indices_matter : env -> bool

val is_impredicative_sort : env -> Sorts.t -> bool

val empty_context : env -> bool

is the local context empty

Context of de Bruijn variables (rel_context)

val nb_rel : env -> int

val push_rel : Constr.rel_declaration -> env -> env

val push_rel_context : Constr.rel_context -> env -> env

val push_rec_types : Constr.rec_declaration -> env -> env

val push_rel_context_val : Constr.rel_declaration -> rel_context_val -> rel_context_val

val set_rel_context_val : rel_context_val -> env -> env

val empty_rel_context_val : rel_context_val

val lookup_rel : int -> env -> Constr.rel_declaration

Looks up in the context of local vars referred by indice (rel_context) raises Not_found if the index points out of the context

val lookup_rel_ctxt : int -> rel_context_val -> Constr.rel_declaration

val evaluable_rel : int -> env -> bool

val env_of_rel : int -> env -> env

Recurrence on rel_context

val fold_rel_context : (env -> Constr.rel_declaration -> 'a -> 'a) -> env -> init:'a -> 'a

Context of variables (section variables and goal assumptions)

type var_status =

| SecVar

| ProofVar

val var_status_eq : var_status -> var_status -> bool

val var_status_ctxt : ?check:bool -> Names.Id.t -> named_context_val -> var_status

val var_status : ?check:bool -> Names.Id.t -> env -> var_status

val section_variables_ctxt : named_context_val -> Names.Id.Set.t

val section_variables : env -> Names.Id.Set.t

val named_context_of_val : named_context_val -> Constr.named_context

val val_of_named_context : (var_status * Constr.named_declaration) list -> named_context_val

val empty_named_context_val : named_context_val

val ids_of_named_context_val : named_context_val -> Names.Id.Set.t

val named_context_of_val_with_status : named_context_val -> (var_status * Constr.named_declaration) list

val map_named_val : (var_status -> Constr.named_declaration -> var_status * Constr.named_declaration) -> named_context_val -> named_context_val

map_named_val f ctxt apply f to the body and the type of each declarations. *** /!\ *** f t must preserve the name

val push_named : var_status -> Constr.named_declaration -> env -> env

val push_named_context : (var_status * Constr.named_declaration) list -> env -> env

val push_named_context_val : var_status -> Constr.named_declaration -> named_context_val -> named_context_val

val mem_named_ctxt : Names.variable -> named_context_val -> bool

val mem_named : Names.variable -> env -> bool

Looks up in the context of local vars referred by names (named_context) raises Not_found if the Id.t is not found

val lookup_named : Names.variable -> env -> Constr.named_declaration

val lookup_named_ctxt : Names.variable -> named_context_val -> Constr.named_declaration

val lookup_named_ctxt_pos : int -> named_context_val -> Constr.named_declaration

val nb_named : named_context_val -> int

val evaluable_named : Names.variable -> env -> bool

val named_type : Names.variable -> env -> Constr.types

val named_body : Names.variable -> env -> Constr.constr option

Recurrence on named_context: older declarations processed first

val fold_named_context_val : (named_context_val -> var_status -> Constr.named_declaration -> 'a -> 'a) -> named_context_val -> init:'a -> 'a

val fold_named_context : (env -> var_status -> Constr.named_declaration -> 'a -> 'a) -> env -> init:'a -> 'a

val match_named_context_val : named_context_val -> (var_status * Constr.named_declaration * named_context_val) option

val fold_named_context_reverse : ('a -> Constr.named_declaration -> 'a) -> init:'a -> env -> 'a

Recurrence on named_context starting from younger decl

val reset_context : env -> env

This forgets named and rel contexts

val reset_with_named_context : named_context_val -> env -> env

This forgets rel context and sets a new named context

val pop_rel_context : int -> env -> env

This removes the n last declarations from the rel context

val fold_constants : (Names.Constant.t -> Declarations.constant_body -> 'a -> 'a) -> env -> 'a -> 'a

Useful for printing

val fold_inductives : (Names.MutInd.t -> Declarations.mutual_inductive_body -> 'a -> 'a) -> env -> 'a -> 'a

Global constants

Add entries to global environment

val add_constant : Names.Constant.t -> Declarations.constant_body -> env -> env

val add_constant_key : Names.Constant.t -> Declarations.constant_body -> link_info -> env -> env

val lookup_constant : Names.Constant.t -> env -> Declarations.constant_body

Looks up in the context of global constant names raises an anomaly if the required path is not found

val lookup_constant_opt : Names.Constant.t -> env -> Declarations.constant_body option

val lookup_constant_key : Names.Constant.t -> env -> link_info Stdlib.ref * key

val lookup_constant_canonical : Names.Constant.t -> env -> Names.KerName.t

val evaluable_constant : Names.Constant.t -> env -> bool

val constant_relevance : Names.Constant.t -> env -> Sorts.relevance

val mem_constant : Names.Constant.t -> env -> bool

val add_rewrite_rules : (Names.Constant.t * Declarations.machine_rewrite_rule) list -> env -> env

val lookup_rewrite_rules : Names.Constant.t -> env -> Declarations.machine_rewrite_rule list

val polymorphic_constant : Names.Constant.t -> env -> bool

New-style polymorphism

val polymorphic_pconstant : Constr.pconstant -> env -> bool

val type_in_type_constant : Names.Constant.t -> env -> bool

...

constant_value env c raises NotEvaluableConst Opaque if c is opaque, NotEvaluableConst NoBody if it has no body, NotEvaluableConst IsProj if c is a projection, NotEvaluableConst (IsPrimitive p) if c is primitive p and an anomaly if it does not exist in env

type const_evaluation_result =

| NoBody

| Opaque

| IsPrimitive of UVars.Instance.t * CPrimitives.t

| HasRules of UVars.Instance.t * bool * Declarations.machine_rewrite_rule list

exception NotEvaluableConst of const_evaluation_result

val constant_type : env -> Names.Constant.t UVars.puniverses -> Constr.types PConstraints.constrained

val constant_value_and_type : env -> Names.Constant.t UVars.puniverses -> Constr.constr option * Constr.types * PConstraints.t

The universe context associated to the constant, empty if not polymorphic

val constant_context : env -> Names.Constant.t -> UVars.AbstractContext.t

The universe context associated to the constant, empty if not polymorphic

val constant_value_in : env -> Names.Constant.t UVars.puniverses -> Constr.constr

val constant_type_in : env -> Names.Constant.t UVars.puniverses -> Constr.types

val constant_opt_value_in : env -> Names.Constant.t UVars.puniverses -> Constr.constr option

val is_symbol : env -> Names.Constant.t -> bool

val is_primitive : env -> Names.Constant.t -> bool

val get_primitive : env -> Names.Constant.t -> CPrimitives.t option

val is_array_type : env -> Names.Constant.t -> bool

val is_int63_type : env -> Names.Constant.t -> bool

val is_float64_type : env -> Names.Constant.t -> bool

val is_string_type : env -> Names.Constant.t -> bool

val is_primitive_type : env -> Names.Constant.t -> bool

Primitive projections

val lookup_projection : Names.Projection.t -> env -> Sorts.relevance * Constr.types

Checks that the number of parameters is correct.

val get_projection : env -> Names.inductive -> proj_arg:int -> Names.Projection.Repr.t * Sorts.relevance

Anomaly when not a primitive record or invalid proj_arg.

val get_projections : env -> Names.inductive -> (Names.Projection.Repr.t * Sorts.relevance) array option

val projection_repr_label : env -> Names.Projection.Repr.t -> Names.Id.t

val projection_repr_constant : env -> Names.Projection.Repr.t -> Names.Constant.t

val lookup_mind_key : Names.MutInd.t -> env -> link_info Stdlib.ref

Inductive types

val add_mind_key : Names.MutInd.t -> Declarations.mutual_inductive_body -> link_info -> env -> env

val add_mind : Names.MutInd.t -> Declarations.mutual_inductive_body -> env -> env

val lookup_mind : Names.MutInd.t -> env -> Declarations.mutual_inductive_body

Looks up in the context of global inductive names raises an anomaly if the required path is not found

val lookup_mind_canonical : Names.MutInd.t -> env -> Names.KerName.t

Returns the canonical name of the inductive

val mem_mind : Names.MutInd.t -> env -> bool

val ind_relevance : Names.inductive -> env -> Sorts.relevance

val mind_context : env -> Names.MutInd.t -> UVars.AbstractContext.t

The universe context associated to the inductive, empty if not polymorphic

val polymorphic_ind : Names.inductive -> env -> bool

New-style polymorphism

val polymorphic_pind : Constr.pinductive -> env -> bool

val type_in_type_ind : Names.inductive -> env -> bool

val template_polymorphic_ind : Names.inductive -> env -> bool

Old-style polymorphism

val template_polymorphic_pind : Constr.pinductive -> env -> bool

Changes of representation of Case nodes

val expand_arity : Declarations.mind_specif -> Constr.pinductive -> Constr.constr array -> (Names.Name.t, _) Context.pbinder_annot array -> Constr.rel_context

Given an inductive type and its parameters, builds the context of the return clause, including the inductive being eliminated.

The additional binder array is only used to set the names of the context variables, we use the less general type to make it easy to use this function on Case nodes.

val expand_branch_contexts : Declarations.mind_specif -> UVars.Instance.t -> Constr.constr array -> (Names.Name.t Constr.binder_annot array * 'a) array -> Constr.rel_context array

Given an inductive type and its parameters, builds the context of the return clause, including the inductive being eliminated. The additional binder array is only used to set the names of the context variables, we use the less general type to make it easy to use this function on Case nodes.

val instantiate_context : UVars.Instance.t -> Vars.substl -> Names.Name.t Constr.binder_annot array -> Constr.rel_context -> Constr.rel_context

instantiate_context u subst nas ctx applies both u and subst to ctx while replacing names using nas (order reversed). In particular, assumes that ctx and nas have the same length.

Name quotients

module type QS = sig ... end

module type QMapS = sig ... end

module QMap (M : CSig.UMapS) (_ : QS with type t = M.key) : QMapS with type key = M.key

module type QNameS = sig ... end

module QConstant : sig ... end

module QMutInd : sig ... end

module QInd : sig ... end

module QConstruct : sig ... end

module QProjection : sig ... end

module QGlobRef : sig ... end

Modules

val add_modtype : Names.ModPath.t -> Mod_declarations.module_type_body -> env -> env

val shallow_add_module : Names.ModPath.t -> Mod_declarations.module_body -> env -> env

shallow_add_module does not add module components

val lookup_module : Names.ModPath.t -> env -> Mod_declarations.module_body

val lookup_modtype : Names.ModPath.t -> env -> Mod_declarations.module_type_body

Universe constraints

val check_univ_constraints : Univ.UnivConstraints.t -> env -> bool

Check universe constraints are satifiable in the environment.

val check_constraints : PConstraints.t -> env -> bool

Check constraints are satifiable in the environment.

val check_ucontext : UVars.UContext.t -> env -> unit

Check that the sort context of the argument satisfies the well-formedness conditions ensuring the existence of an instance.

val push_context : ?strict:bool -> UVars.UContext.t -> env -> env

push_context ?(strict=false) ctx env pushes the universe context to the environment.

raises UGraph.AlreadyDeclared
if one of the universes is already declared.

val push_context_set : ?strict:bool -> Univ.ContextSet.t -> env -> env

push_context_set ?(strict=false) src ctx env pushes the universe context set to the environment. It does not fail even if one of the universes is already declared.

val push_qualities : Sorts.Quality.Set.t -> env -> env

push_qualities qs env pushes the set of quality variables and constraints in the environment. It fails if a quality variable is already declared.

val merge_elim_constraints : rigid:bool -> Sorts.ElimConstraints.t -> env -> env

merge_elim_constraints ~rigid qcsts env adds the elimination constraints to the graph, rigidly if rigid.

val push_subgraph : Univ.ContextSet.t -> env -> env

push_subgraph univs env adds the universes and constraints in univs to env as push_context_set ~strict:false univs env, and also checks that they do not imply new transitive constraints between pre-existing universes in env.

val set_typing_flags : Declarations.typing_flags -> env -> env

val set_impredicative_set : bool -> env -> env

val set_type_in_type : bool -> env -> env

val set_allow_sprop : bool -> env -> env

val sprop_allowed : env -> bool

val allow_rewrite_rules : env -> env

val rewrite_rules_allowed : env -> bool

val same_flags : Declarations.typing_flags -> Declarations.typing_flags -> bool

val update_typing_flags : ?typing_flags:Declarations.typing_flags -> env -> env

update_typing_flags ?typing_flags may update env with optional typing flags

val universes_of_global : env -> Names.GlobRef.t -> UVars.AbstractContext.t

Sets of referred section variables

global_vars_set env c returns the list of id's occurring either directly as Var id in c or indirectly as a section variable dependent in a global reference occurring in c

val global_vars_set : env -> Constr.constr -> Names.Id.Set.t

val vars_of_global : env -> Names.GlobRef.t -> Names.Id.Set.t

val really_needed : env -> Names.Id.Set.t -> Names.Id.Set.t

closure of the input id set w.r.t. dependency

val keep_hyps : env -> Names.Id.Set.t -> Constr.named_context

like really_needed but computes a well ordered named context

Unsafe judgments.

We introduce here the pre-type of judgments, which is actually only a datatype to store a term with its type and the type of its type.

type ('constr, 'types) punsafe_judgment = {

uj_val : 'constr;

uj_type : 'types;

}

val on_judgment : ('a -> 'b) -> ('a, 'a) punsafe_judgment -> ('b, 'b) punsafe_judgment

val on_judgment_value : ('c -> 'c) -> ('c, 't) punsafe_judgment -> ('c, 't) punsafe_judgment

val on_judgment_type : ('t -> 't) -> ('c, 't) punsafe_judgment -> ('c, 't) punsafe_judgment

type unsafe_judgment = (Constr.constr, Constr.types) punsafe_judgment

val make_judge : 'constr -> 'types -> ('constr, 'types) punsafe_judgment

val j_val : ('constr, 'types) punsafe_judgment -> 'constr

val j_type : ('constr, 'types) punsafe_judgment -> 'types

type ('types, 'sorts) punsafe_type_judgment = {

utj_val : 'types;

utj_type : 'sorts;

}

type unsafe_type_judgment = (Constr.types, Sorts.t) punsafe_type_judgment

exception Hyp_not_found

val apply_to_hyp : named_context_val -> Names.variable -> (Constr.named_context -> var_status -> Constr.named_declaration -> Constr.named_context -> var_status * Constr.named_declaration) -> named_context_val

apply_to_hyp sign id f split sign into tail::(id,_,_)::head and return tail::(f head (id,_,_) (rev tail))::head. the value associated to id should not change

val remove_hyps : Names.Id.Set.t -> (var_status -> Constr.named_declaration -> var_status * Constr.named_declaration) -> named_context_val -> named_context_val

val is_polymorphic : env -> Names.GlobRef.t -> bool

val is_template_polymorphic : env -> Names.GlobRef.t -> bool

val is_type_in_type : env -> Names.GlobRef.t -> bool

val ind_ignores_elim_constraints : env -> Names.inductive -> bool

VM and native

val vm_library : env -> Vmlibrary.t

val set_vm_library : Vmlibrary.t -> env -> env

val link_vm_library : Vmlibrary.on_disk -> env -> env

val lookup_vm_code : Vmlibrary.index -> env -> Vmemitcodes.to_patch

val no_link_info : link_info

Native compiler

val set_retroknowledge : env -> Retroknowledge.retroknowledge -> env

Primitives

val retroknowledge : env -> Retroknowledge.retroknowledge

Dependency analysis

val constant_depends_on : env -> Names.Constant.t -> Names.Constant.t -> bool

constant_depends_on c1 c2 is true when c2 appears in the transitive set of constants reachable from the body of c1. Axioms, opaque definitions, and primitives have no body and thus no dependencies.

Internals

module Internal : sig ... end