Module Esubst.Internal

Debugging utilities

val pp_lift : lift -> Pp.t

If Γ ⊢ lift : Δ, pp_lift lift prints Γ as a list of whether the corresponding element appears in Δ

type 'a or_rel =

| REL of int

| VAL of int * 'a

val repr : 'a subs -> 'a or_rel list * int

High-level representation of a substitution. The first component is a list that associates a value to an index, and the second component is the relocation shift that must be applied to any variable pointing outside of the substitution.