Module Extraction_plugin

module Common : sig ... end

By default, in module Format, you can do horizontal placing of blocks even if they include newlines, as long as the number of chars in the blocks are less that a line length. To avoid this awkward situation, we attach a big virtual size to fnl newlines.

module Extract_env : sig ... end

module Extraction : sig ... end

module G_extraction : sig ... end

module Haskell : sig ... end

module Json : sig ... end

module Miniml : sig ... end

module Mlutil : sig ... end

module Modutil : sig ... end

module Ocaml : sig ... end

module Scheme : sig ... end

module Table : sig ... end