Module Hints

General functions.

exception Bound

val decompose_app_bound : Environ.env -> Evd.evar_map -> EConstr.constr -> Names.GlobRef.t * EConstr.constr array

type debug =

| Debug

| Info

| Off

val secvars_of_hyps : Environ.named_context_val -> Names.Id.Pred.t

val empty_hint_info : 'a Typeclasses.hint_info_gen

val hint_cat : Libobject.category

Pre-created hint databases

type hint

type hint_ast =

| Res_pf of hint

| ERes_pf of hint

| Give_exact of hint

| Res_pf_THEN_trivial_fail of hint

| Unfold_nth of Evaluable.t

| Extern of Pattern.constr_pattern option * Gentactic.glob_generic_tactic

val hint_as_term : hint -> UnivGen.sort_context_set option * EConstr.constr

type 'a hints_path_atom_gen =

| PathHints of 'a list

| PathAny

type hints_path_atom = Names.GlobRef.t hints_path_atom_gen

type hint_db_name = string

module FullHint : sig ... end

The head may not be bound.

type hint_mode =

| ModeInput

| ModeFrozen

| ModeNoHeadEvar

| ModeOutput

type 'a hints_transparency_target =

| HintsVariables

| HintsConstants

| HintsProjections

| HintsReferences of 'a list

type 'a hints_path_gen =

| PathAtom of 'a hints_path_atom_gen

| PathStar of 'a hints_path_gen

| PathSeq of 'a hints_path_gen * 'a hints_path_gen

| PathOr of 'a hints_path_gen * 'a hints_path_gen

| PathEmpty

| PathEpsilon

type pre_hints_path = Libnames.qualid hints_path_gen

type hints_path = Names.GlobRef.t hints_path_gen

val path_matches_epsilon : hints_path -> bool

val path_derivate : Environ.env -> hints_path -> Names.GlobRef.t option -> hints_path

val pp_hints_path_gen : ('a -> Pp.t) -> 'a hints_path_gen -> Pp.t

val parse_mode : string -> hint_mode

val parse_modes : string -> hint_mode list

val string_of_mode : hint_mode -> string

val pp_hint_mode : hint_mode -> Pp.t

val glob_hints_path : pre_hints_path -> hints_path

type mode_match =

| NoMode

| WithMode of Evarsolve.AllowedEvars.t

type 'a with_mode =

| ModeMatch of mode_match * 'a

| ModeMismatch

module Modes : sig ... end

module Hint_db : sig ... end

type hint_db = Hint_db.t

type hnf = bool

type hints_entry =

| HintsResolveEntry of (Typeclasses.hint_info * hnf * Names.GlobRef.t) list

| HintsImmediateEntry of Names.GlobRef.t list

| HintsCutEntry of hints_path

| HintsUnfoldEntry of Evaluable.t list

| HintsTransparencyEntry of Evaluable.t hints_transparency_target * bool

| HintsModeEntry of Names.GlobRef.t * hint_mode list

| HintsExternEntry of Typeclasses.hint_info * Gentactic.glob_generic_tactic

val searchtable_map : hint_db_name -> hint_db

val searchtable_add : (hint_db_name * hint_db) -> unit

type hint_locality = Libobject.locality =

| Local

| Export

| SuperGlobal

create_hint_db local name st use_dn. st is a transparency state for unification using this db use_dn switches the use of the discrimination net for all hints and patterns.

val create_hint_db : bool -> hint_db_name -> TransparentState.t -> bool -> unit

val remove_hints : locality:hint_locality -> hint_db_name list -> Names.GlobRef.t list -> unit

val current_db_names : unit -> Util.String.Set.t

val current_pure_db : unit -> hint_db list

val add_hints : locality:hint_locality -> hint_db_name list -> hints_entry -> unit

A constr which is Hint'ed will be:

(1) used as an Exact, if it does not start with a product

(2) used as an Apply, if its HNF starts with a product, and has no missing arguments.

(3) used as an EApply, if its HNF starts with a product, and has missing arguments.

val push_resolves : Environ.env -> Evd.evar_map -> Names.GlobRef.t -> Hint_db.t -> Hint_db.t

push_resolve_hyp hname htyp db. used to add an hypothesis to the local hint database; Never raises a user exception; If the hyp cannot be used as a Hint, it is not added.

val push_resolve_hyp : Environ.env -> Evd.evar_map -> Names.Id.t -> Hint_db.t -> Hint_db.t

Create a Hint database from the pairs (name, constr). Useful to take the current goal hypotheses as hints; Boolean tells if lemmas with evars are allowed

val make_local_hint_db : Environ.env -> Evd.evar_map -> ?ts:TransparentState.t -> bool -> Names.GlobRef.t list -> hint_db

val make_db_list : hint_db_name list -> hint_db list

val fresh_hint : Environ.env -> Evd.evar_map -> hint -> Evd.evar_map * EConstr.constr

val hint_res_pf : ?with_evars:bool -> ?with_classes:bool -> ?flags:Unification.unify_flags -> hint -> unit Proofview.tactic

Printing hints

val pr_searchtable : Environ.env -> Evd.evar_map -> Pp.t

val pr_applicable_hint : Proof.t -> Pp.t

val pr_hint_ref : Environ.env -> Evd.evar_map -> Names.GlobRef.t -> Pp.t

val pr_hint_db_by_name : Environ.env -> Evd.evar_map -> hint_db_name -> Pp.t

val pr_hint_db_env : Environ.env -> Evd.evar_map -> Hint_db.t -> Pp.t