Module Libnames

val dirpath_of_string : string -> Names.DirPath.t

Dirpaths

val pop_dirpath : Names.DirPath.t -> Names.DirPath.t

Pop the suffix of a DirPath.t. Raises a Failure for an empty path

val pop_dirpath_n : int -> Names.DirPath.t -> Names.DirPath.t

Pop the suffix n times

val split_dirpath : Names.DirPath.t -> Names.DirPath.t * Names.Id.t

Immediate prefix and basename of a DirPath.t. May raise Failure

val add_dirpath_suffix : Names.DirPath.t -> Names.Id.t -> Names.DirPath.t

val append_dirpath : Names.DirPath.t -> Names.DirPath.t -> Names.DirPath.t

val drop_dirpath_prefix : Names.DirPath.t -> Names.DirPath.t -> Names.DirPath.t

val is_dirpath_prefix_of : Names.DirPath.t -> Names.DirPath.t -> bool

val is_dirpath_suffix_of : Names.DirPath.t -> Names.DirPath.t -> bool

type full_path

Full paths are absolute paths of declarations

val eq_full_path : full_path -> full_path -> bool

val compare_full_path : full_path -> full_path -> int

val make_path : Names.DirPath.t -> Names.Id.t -> full_path

Constructors of full_path

val add_path_suffix : full_path -> Names.Id.t -> full_path

val append_path : full_path -> Names.DirPath.t -> full_path

val repr_path : full_path -> Names.DirPath.t * Names.Id.t

Destructors of full_path

val path_pop_n_suffixes : int -> full_path -> full_path

path_pop_n_suffixes n p removes the last n elements of p. Raises Failure if p is not long enough.

val path_pop_suffix : full_path -> full_path

path_pop_suffix p is path_pop_n_suffixes 1 p.

val dirpath : full_path -> Names.DirPath.t

The prefix of the path

deprecated (9.1) Compose [dirpath_of_path] and [pop_dirpath]

val basename : full_path -> Names.Id.t

val dirpath_of_path : full_path -> Names.DirPath.t

The full path as a DirPath.t.

val is_path_prefix_of : full_path -> Names.DirPath.t -> bool

val path_of_string : string -> full_path

Parsing and printing of section path as "root.module.id"

val string_of_path : full_path -> string

val pr_path : full_path -> Pp.t

module Spmap : CSig.MapS with type key = full_path

...

A qualid is a partially qualified ident; it includes fully qualified names (= absolute names) and all intermediate partial qualifications of absolute names, including single identifiers. The qualid are used to access the name table.

type qualid_r

type qualid = qualid_r CAst.t

val make_qualid : ?loc:Loc.t -> Names.DirPath.t -> Names.Id.t -> qualid

val repr_qualid : qualid -> Names.DirPath.t * Names.Id.t

val qualid_eq : qualid -> qualid -> bool

val qualid_compare : qualid -> qualid -> int

val is_qualid_suffix_of_full_path : qualid -> full_path -> bool

val pr_qualid : qualid -> Pp.t

val string_of_qualid : qualid -> string

val qualid_of_string : ?loc:Loc.t -> string -> qualid

Turns an absolute name, a dirpath, or an Id.t into a qualified name denoting the same name

val qualid_of_path : ?loc:Loc.t -> full_path -> qualid

val qualid_of_dirpath : ?loc:Loc.t -> Names.DirPath.t -> qualid

val qualid_of_ident : ?loc:Loc.t -> Names.Id.t -> qualid

val qualid_of_lident : Names.lident -> qualid

val qualid_is_ident : qualid -> bool

val qualid_path : qualid -> Names.DirPath.t

val qualid_basename : qualid -> Names.Id.t

val idset_mem_qualid : qualid -> Names.Id.Set.t -> bool

false when the qualid is not an ident

...

val rocq_init_root : Names.Id.t

This is the root of the rocq-core library

val rocq_init_string : string

val default_root_prefix : Names.DirPath.t

This is the default root prefix for developments which doesn't mention a root

val dummy_full_path : full_path

For uninitialized data, cf DirPath.dummy