Module Ltac2_plugin.Tac2core

val throw : ?info:Exninfo.info -> exn -> 'a Proofview.tactic

val pf_apply : ?catch_exceptions:bool -> (Environ.env -> Evd.evar_map -> 'a Proofview.tactic) -> 'a Proofview.tactic

catch_exceptions default false

val wrap_exceptions : ?passthrough:bool -> (unit -> 'a Proofview.tactic) -> 'a Proofview.tactic

Adds ltac2 backtrace. With passthrough:false (default), acts like Proofview.wrap_exceptions + Ltac2 backtrace handling.

val constr_flags : Pretyping.inference_flags

val open_constr_no_classes_flags : Pretyping.inference_flags

val preterm_flags : Pretyping.inference_flags

module type MapType = sig ... end

type ('a, 'set, 'map) map_tag

val map_tag_eq : ('a, 'set1, 'map1) map_tag -> ('b, 'set2, 'map2) map_tag -> ('a * 'set1 * 'map1, 'b * 'set2 * 'map2) Util.eq option

type any_map_tag =

| Any : (_, _, _) map_tag -> any_map_tag

type tagged_set =

| TaggedSet : (_, 'set, _) map_tag * 'set -> tagged_set

type tagged_map =

| TaggedMap : (_, _, 'map) map_tag * 'map -> tagged_map

val map_tag_repr : any_map_tag Tac2ffi.repr

val set_repr : tagged_set Tac2ffi.repr

val map_repr : tagged_map Tac2ffi.repr

val tag_set : (_, 'set, _) map_tag -> 'set -> Tac2val.valexpr

val tag_map : (_, _, 'map) map_tag -> 'map -> Tac2val.valexpr

val register_map : ?plugin:string -> tag_name:string -> (module MapType with type valmap = 'map and type S.elt = 'a and type S.t = 'set) -> ('a, 'set, 'map) map_tag

Register a type on which we can use finite sets and maps. tag_name is the name used for the external to make the Ltac2.FSet.Tags.tag available.

val get_map : ('a, 'set, 'map) map_tag -> (module MapType with type valmap = 'map and type S.elt = 'a and type S.t = 'set)

Default registered maps

val ident_map_tag : (Names.Id.t, Names.Id.Set.t, Tac2val.valexpr Names.Id.Map.t) map_tag

val int_map_tag : (int, Int.Set.t, Tac2val.valexpr Int.Map.t) map_tag

val string_map_tag : (string, CString.Set.t, Tac2val.valexpr CString.Map.t) map_tag

val inductive_map_tag : (Names.inductive, Names.Indset_env.t, Tac2val.valexpr Names.Indmap_env.t) map_tag

val constructor_map_tag : (Names.constructor, Names.Constrset_env.t, Tac2val.valexpr Names.Constrmap_env.t) map_tag

val constant_map_tag : (Names.Constant.t, Names.Cset_env.t, Tac2val.valexpr Names.Cmap_env.t) map_tag