Module Mltop

Toplevel management

module PluginSpec : sig ... end

Coq plugins are identified by their OCaml library name (in the Findlib sense)

type toplevel = {

load_plugin : PluginSpec.t -> unit;
(*
Load a findlib library, given by public name
*)

load_module : string -> unit;
(*
Load a cmxs / cmo module, used by the native compiler to load objects
*)

add_dir : string -> unit;
(*
Adds a dir to the module search path
*)

ml_loop : ?init_file:string -> unit -> unit;
(*
Run the OCaml toplevel with given addtitional initialisation file
*)

}

val set_top : toplevel -> unit

Sets and initializes a toplevel (if any)

val load_module : string -> unit

Low level module loading, for the native compiler and similar users.

val remove : unit -> unit

Removes the toplevel (if any)

val is_ocaml_top : unit -> bool

Tests if an Ocaml toplevel runs under Coq

val ocaml_toploop : ?init_file:string -> unit -> unit

Starts the Ocaml toplevel loop

ML Dynlink

val add_ml_dir : string -> unit

Adds a dir to the plugin search path, this also extends OCamlfind's search path

val has_dynlink : bool

Tests if we can load ML files

Initialization functions

val add_init_function : string -> (unit -> unit) -> unit

Declare a initialization function. The initialization function is called in Declare ML Module, including reruns after backtracking over it (either interactive backtrack, module closing backtrack, Require of a file with Declare ML Module).

val declare_cache_obj : (unit -> unit) -> string -> unit

Register a callback that will be called when the module is declared with the Declare ML Module command. This is useful to define Coq objects at that time only. Several functions can be defined for one module; they will be called in the order of declaration, and after the ML module has been properly initialized.

Unlike the init functions it does not run after closing a module or Requiring a file which contains the Declare ML Module. This allows to have effects which depend on the module when command was run in, eg add a named libobject which will use it for the prefix.

The callback runs in the synterp phase, use declare_cache_obj_full if you also need to interact with Interp state.

type cache_obj =

| CacheObj : {

synterp : unit -> 'a;

interp : 'a -> unit;

} -> cache_obj

val interp_only_obj : (unit -> unit) -> cache_obj

val declare_cache_obj_full : cache_obj -> string -> unit

Register a callback with an interp phase.

Declaring modules

type interp_fun

val run_interp_fun : interp_fun -> unit

val declare_ml_modules : Vernacexpr.locality_flag -> string list -> interp_fun

Implementation of the Declare ML Module vernacular command.

Utilities

val print_ml_modules : unit -> Pp.t

val print_gc : unit -> Pp.t