Module Mltop.PluginSpec

Coq plugins are identified by their OCaml library name (in the Findlib sense)

type t

A plugin is identified by its canonical library name, such as rocq-runtime.plugins.ltac

val to_package : t -> string

to_package p returns the findlib name of the package

val pp : t -> string