Module Modops

Various operations on modules and module types

Functors

val is_functor : ('ty, 'a) Declarations.functorize -> bool

val destr_functor : ('ty, 'a) Declarations.functorize -> Names.MBId.t * 'ty * ('ty, 'a) Declarations.functorize

val destr_nofunctor : Names.ModPath.t -> ('ty, 'a) Declarations.functorize -> 'a

val check_modpath_equiv : Environ.env -> Names.ModPath.t -> Names.ModPath.t -> unit

val annotate_module_expression : Declarations.module_expression -> Mod_declarations.module_signature -> (Mod_declarations.module_type_body, (Constr.constr * UVars.AbstractContext.t option) Declarations.module_alg_expr) Declarations.functorize

val annotate_struct_body : Mod_declarations.structure_body -> Mod_declarations.module_signature -> Mod_declarations.module_signature

Substitutions

val subst_signature : Mod_subst.substitution -> Names.ModPath.t -> Mod_declarations.module_signature -> Mod_declarations.module_signature

val subst_structure : Mod_subst.substitution -> Names.ModPath.t -> Mod_declarations.structure_body -> Mod_declarations.structure_body

Adding to an environment

val add_structure : Names.ModPath.t -> Mod_declarations.structure_body -> Mod_subst.delta_resolver -> Environ.env -> Environ.env

val add_module : Names.ModPath.t -> Mod_declarations.module_body -> Environ.env -> Environ.env

adds a module and its components, but not the constraints

val add_linked_module : Names.ModPath.t -> Mod_declarations.module_body -> Environ.link_info -> Environ.env -> Environ.env

same as add_module, but for a module whose native code has been linked by the native compiler. The linking information is updated.

val add_module_parameter : Names.MBId.t -> Mod_declarations.module_type_body -> Environ.env -> Environ.env

add an abstract module parameter to the environment

val add_retroknowledge : Retroknowledge.action list -> Environ.env -> Environ.env

Strengthening

val strengthen : Mod_declarations.module_type_body -> Names.ModPath.t -> Mod_declarations.module_type_body

val strengthen_and_subst_module_body : Names.ModPath.t -> Mod_declarations.module_body -> Names.ModPath.t -> bool -> Mod_declarations.module_body

val subst_modtype_signature_and_resolver : Names.ModPath.t -> Names.ModPath.t -> Mod_declarations.module_signature -> Mod_subst.delta_resolver -> Mod_declarations.module_signature * Mod_subst.delta_resolver

Building map of constants to inline

val inline_delta_resolver : Environ.env -> Entries.inline -> Names.ModPath.t -> Names.MBId.t -> Mod_declarations.module_type_body -> Mod_subst.delta_resolver -> Mod_subst.delta_resolver

Cleaning a module expression from bounded parts

For instance: functor(X:T)->struct module M:=X end) becomes: functor(X:T)->struct module M:=<content of T> end)

val clean_bounded_mod_expr : Mod_declarations.module_signature -> Mod_declarations.module_signature

Errors

type signature_mismatch_error =

| InductiveFieldExpected of Declarations.mutual_inductive_body

| DefinitionFieldExpected

| ModuleFieldExpected

| ModuleTypeFieldExpected

| NotConvertibleInductiveField of Names.Id.t * (Environ.env * Constr.types * Constr.types) option

| NotConvertibleConstructorField of Names.Id.t * (Environ.env * Constr.types * Constr.types) option

| NotConvertibleBodyField of (Environ.env * Constr.constr * Constr.constr) option

| NotConvertibleTypeField of Environ.env * Constr.types * Constr.types

| CumulativeStatusExpected of bool

| PolymorphicStatusExpected of bool

| NotSameConstructorNamesField of Names.Id.t array * Names.Id.t array

| NotSameInductiveNameInBlockField of Names.Id.t * Names.Id.t

| FiniteInductiveFieldExpected of bool

| InductiveNumbersFieldExpected of {

got : int;

expected : int;

}

| InductiveParams of {

env : Environ.env;

got : Constr.rel_context;

expected : Constr.rel_context;

}

| RecordFieldExpected of bool

| RecordProjectionsExpected of {

expected : Names.Name.t list;

got : Names.Name.t list;

}

| NotEqualInductiveAliases of Names.MutInd.t * Names.MutInd.t

| IncompatibleUniverses of {

err : UGraph.univ_inconsistency;

env : Environ.env;

t1 : Constr.types;

t2 : Constr.types;

}

| IncompatibleQualities of {

err : QGraph.elimination_error;

env : Environ.env;

t1 : Constr.types;

t2 : Constr.types;

}

| IncompatiblePolymorphism of Environ.env * Constr.types * Constr.types

| IncompatibleUnivConstraints of {

env : Environ.env;

got : UVars.AbstractContext.t;

expect : UVars.AbstractContext.t;

}

| IncompatibleVariance

| NoRewriteRulesSubtyping

type with_constraint_error =

| WithSignatureMismatch of signature_mismatch_error

| WithCannotConstrainPrimitive

| WithCannotConstrainSymbol

type subtyping_trace_elt =

| Submodule of Names.Id.t

| FunctorArgument of int

type module_typing_error =

| SignatureMismatch of subtyping_trace_elt list * Names.Id.t * signature_mismatch_error

| LabelAlreadyDeclared of Names.Id.t

| NotAFunctor

| IsAFunctor of Names.ModPath.t

| IncompatibleModuleTypes of Mod_declarations.module_type_body * Mod_declarations.module_type_body

| NotEqualModulePaths of Names.ModPath.t * Names.ModPath.t

| NoSuchLabel of Names.Id.t * Names.ModPath.t

| NotAModuleLabel of Names.Id.t

| NotAConstant of Names.Id.t

| IncorrectWithConstraint of Names.Id.t * with_constraint_error

| GenerativeModuleExpected of Names.Id.t

| LabelMissing of Names.Id.t * string

| IncludeRestrictedFunctor of Names.ModPath.t

exception ModuleTypingError of module_typing_error

val error_existing_label : Names.Id.t -> 'a

val error_incompatible_modtypes : Mod_declarations.module_type_body -> Mod_declarations.module_type_body -> 'a

val error_signature_mismatch : subtyping_trace_elt list -> Names.Id.t -> signature_mismatch_error -> 'a

val error_no_such_label : Names.Id.t -> Names.ModPath.t -> 'a

val error_not_a_module_label : Names.Id.t -> 'a

val error_not_a_constant : Names.Id.t -> 'a

val error_incorrect_with_constraint : Names.Id.t -> with_constraint_error -> 'a

val error_generative_module_expected : Names.Id.t -> 'a

val error_no_such_label_sub : Names.Id.t -> string -> 'a

val error_include_restricted_functor : Names.ModPath.t -> 'a