Module Nameops.Subscript

type t

Abstract datatype of subscripts. Isomorphic to a string of digits.

val zero : t

Empty subscript

val succ : t -> t

Guarantees that x < succ x, but succ x might not be the smallest element strictly above x, generally it does not exist. Example mappings: "" ↦ "0" "0" ↦ "1" "00" ↦ "01" "1" ↦ "2" "01" ↦ "02" "9" ↦ "10" "09" ↦ "10" "99" ↦ "100"

val compare : t -> t -> int

Well-founded order.

val equal : t -> t -> bool