Module NumTok.UnsignedNat

type t

val of_string : string -> t

Convert from a non-empty sequence of digits (which may contain "_") (or hexdigits when starting with "0x" or "0X")

val to_string : t -> string

Convert to a non-empty sequence of digit that does not contain "_" (or hexdigits, starting with "0x", all hexdigits are lower case)

val sprint : t -> string

val print : t -> Pp.t

sprint and print returns the number as it was parsed, for printing

val classify : t -> num_class

val compare : t -> t -> int