Module Extern.Records

type t

val make_raw : t -> t

val current : unit -> t

val current_ignore_raw : unit -> t

val use_record_syntax : t -> Names.inductive -> bool

Tells whether record syntax should be used for some record. If the inductive is not a record, the result is meaningless.