Module Proof

type t

type data = private {

sigma : Evd.evar_map;
(*
A representation of the evar_map EJGA wouldn't it better to just return the proofview?
*)

goals : Evar.t list;
(*
Focused goals
*)

entry : Proofview.entry;
(*
Entry for the proofview
*)

stack : (Evar.t list * Evar.t list) list;
(*
A representation of the focus stack
*)

name : Names.Id.t;
(*
The name of the theorem whose proof is being constructed
*)

poly : PolyFlags.t;
(*
Universe polymorphism
*)

}

val data : t -> data

val start : name:Names.Id.t -> poly:PolyFlags.t -> ?typing_flags:Declarations.typing_flags -> Evd.evar_map -> (Environ.env * EConstr.types) list -> t

val dependent_start : name:Names.Id.t -> poly:PolyFlags.t -> ?typing_flags:Declarations.typing_flags -> Proofview.telescope -> t

val is_done : t -> bool

val partial_proof : t -> EConstr.constr list

val compact : t -> t

val update_sigma_univs : UGraph.t -> t -> t

update_sigma_univs lifts UState.update_sigma_univs to the proof

type 'a focus_kind

val new_focus_kind : string -> 'a focus_kind

type 'a focus_condition

val no_cond : ?loose_end:bool -> 'a focus_kind -> 'a focus_condition

val done_cond : ?loose_end:bool -> 'a focus_kind -> 'a focus_condition

val focus : 'a focus_condition -> 'a -> int -> t -> t

val focus_id : 'a focus_condition -> 'a -> Libnames.qualid -> t -> t

exception FullyUnfocused

exception CannotUnfocusThisWay

exception NoSuchGoals of int * int

exception NoSuchGoal of Libnames.qualid option

val unfocus : 'a focus_kind -> t -> unit -> t

val unfocused : t -> bool

val unfocus_all : t -> t

Unfocus everything (fail if no allowed).

exception NoSuchFocus

val get_at_focus : 'a focus_kind -> t -> 'a

val is_last_focus : 'a focus_kind -> t -> bool

val no_focused_goal : t -> bool

val run_tactic : Environ.env -> 'a Proofview.tactic -> t -> t * (Environ.env * bool * Proofview_monad.Info.tree) * 'a

val maximal_unfocus : 'a focus_kind -> t -> t

val unshelve : t -> t

val goal_uid : Evar.t -> string

val pr_proof : t -> Pp.t

val background_subgoals : t -> Evar.t list

val all_goals : t -> Evar.Set.t

solve (select_nth n) tac applies tactic tac to the nth subgoal of the current focused proof. solve SelectAll tac applies tac to all subgoals.

val solve : ?with_end_tac:unit Proofview.tactic -> Environ.env -> Goal_select.t -> int option -> unit Proofview.tactic -> t -> t * bool

val use_unification_heuristics : unit -> bool

Option telling if unification heuristics should be used.

exception SuggestNoSuchGoals of int * t

val get_goal_context_gen : t -> int -> Evd.evar_map * Environ.env

Helpers to obtain proof state when in an interactive proof

val get_proof_context : t -> Evd.evar_map * Environ.env

get_proof_context () gets the goal context for the first subgoal of the proof

val purge_side_effects : t -> t * Evd.side_effects

val set_used_variables : Environ.env -> kept:Names.Id.Set.t -> t -> t

set_used_variables env kept p clears variables not in kept from the proof terms, with an error if this is impossible (eg "#refine" with the refine term using a forbidden variable).

exception ProofUsingClearDependency of Environ.env * Evd.evar_map * Names.Id.t * Evarutil.clear_dependency_error * Names.GlobRef.t option

Exposed for printing