Module Redexpr

Interpretation layer of redexprs such as hnf, cbv, etc.

type 'l user_red_expr

User-defined reductions

val subst_user_red_expr : Genarg.glevel user_red_expr Gensubst.subst_fun

val eval_user_red_expr : Genarg.glevel user_red_expr -> Reductionops.e_reduction_function * Constr.cast_kind

val pr_raw_user_red_expr : Environ.env -> Evd.evar_map -> Genarg.rlevel user_red_expr -> Pp.t

val pr_glob_user_red_expr : Environ.env -> Evd.evar_map -> Genarg.glevel user_red_expr -> Pp.t

type raw_red_expr = Genarg.rlevel user_red_expr Genredexpr.raw_red_expr

type glob_red_expr = Genarg.glevel user_red_expr Genredexpr.glob_red_expr

type red_expr = (EConstr.constr, Evaluable.t, Pattern.constr_pattern, int, Genarg.glevel user_red_expr) Genredexpr.red_expr_gen

type red_expr_val

val out_occurrences : Locus.occurrences_expr -> Locus.occurrences

val out_with_occurrences : 'a Locus.with_occurrences_expr -> 'a Locus.with_occurrences

val eval_red_expr : Environ.env -> red_expr -> red_expr_val

val reduction_of_red_expr_val : ?occs:(Locus.occurrences * int) -> red_expr_val -> Reductionops.e_reduction_function * Constr.cast_kind

val reduction_of_red_expr : Environ.env -> red_expr -> Reductionops.e_reduction_function * Constr.cast_kind

Composition of reduction_of_red_expr_val with eval_red_expr

true if we should use the vm to verify the reduction

val declare_reduction : string -> Reductionops.reduction_function -> unit

Adding a custom reduction (function to be use at the ML level) NB: the effect is permanent.

val declare_red_expr : bool -> string -> red_expr -> unit

Adding a custom reduction (function to be called a vernac command). The boolean flag is the locality.

Opaque and Transparent commands.

val set_strategy : bool -> (Conv_oracle.level * Evaluable.t list) list -> unit

Sets the expansion strategy of a constant. When the boolean is true, the effect is non-synchronous (i.e. it does not survive section and module closure).

val cbv_vm : Reductionops.reduction_function

call by value normalisation function using the virtual machine

val subst_red_expr : Mod_subst.substitution -> red_expr -> red_expr

subst_red_expr sub c performs the substitution sub on all kernel names appearing in c

val wit_red_expr : (raw_red_expr, glob_red_expr, red_expr) Genarg.genarg_type

module Intern : sig ... end

module Interp : sig ... end

val interp_redexp_no_ltac : Environ.env -> Evd.evar_map -> raw_red_expr -> Evd.evar_map * red_expr

module User : sig ... end