Parameter MakeObservable.Obs

type value

An arbitrary type, does not need to be marshallable.

val stage : Stage.t

val local : bool

Whether the list of active observers is process-local. Unlike ref ~local this doesn't matter for marshalling since observers apply indirection to be always marshallable.

val name : string

The name of the summary.