Module Tacmach

Operations for handling terms under a local typing context.

Variants of Tacmach functions built with the new proof engine

val pf_apply : (Environ.env -> Evd.evar_map -> 'a) -> Proofview.Goal.t -> 'a

deprecated (9.2) Use Proofview.Goal.{env, sigma}

val project : Proofview.Goal.t -> Evd.evar_map

deprecated (9.2) Use Proofview.Goal.sigma

val pf_env : Proofview.Goal.t -> Environ.env

deprecated (9.2) Use Proofview.Goal.env

val pf_concl : Proofview.Goal.t -> EConstr.types

deprecated (9.2) Use Proofview.Goal.concl

val pf_get_type_of : Proofview.Goal.t -> EConstr.constr -> EConstr.types

This function does no type inference and expects an already well-typed term. It recomputes its type in the fastest way possible (no conversion is ever involved)

deprecated (9.2) Use Retyping.get_type_of

val pf_type_of : Proofview.Goal.t -> EConstr.constr -> Evd.evar_map * EConstr.types

This function entirely type-checks the term and computes its type and the implied universe constraints.

deprecated (9.2) Use Typing.type_of

val pf_conv_x : Proofview.Goal.t -> EConstr.t -> EConstr.t -> bool

deprecated (9.2) Use Reductionops.is_conv

val pf_get_new_id : Names.Id.t -> Proofview.Goal.t -> Names.Id.t

val pf_ids_of_hyps : Proofview.Goal.t -> Names.Id.t list

val pf_ids_set_of_hyps : Proofview.Goal.t -> Names.Id.Set.t

val pf_hyps_types : Proofview.Goal.t -> (Names.Id.t * EConstr.types) list

deprecated (9.2) Use EConstr.named_context

val pf_get_hyp : Names.Id.t -> Proofview.Goal.t -> EConstr.named_declaration

val pf_get_hyp_typ : Names.Id.t -> Proofview.Goal.t -> EConstr.types

val pf_last_hyp : Proofview.Goal.t -> EConstr.named_declaration

deprecated (9.2) Use EConstr.named_context

val pf_nf_concl : Proofview.Goal.t -> EConstr.types

deprecated (9.2) Use Reductionops.nf_evar

val pf_hnf_constr : Proofview.Goal.t -> EConstr.constr -> EConstr.types

deprecated (9.2) Use Tacred.hnf_constr

val pf_hnf_type_of : Proofview.Goal.t -> EConstr.constr -> EConstr.types

deprecated (9.2) Use Reductionops.whd_all and Retyping.get_type_of

val pf_compute : Proofview.Goal.t -> EConstr.constr -> EConstr.constr

deprecated (9.2) Use Tacred.pf_compute

val pf_whd_compute : Proofview.Goal.t -> EConstr.constr -> EConstr.constr

deprecated (9.2) Use Tacred.whd_compute

val pf_nf_evar : Proofview.Goal.t -> EConstr.constr -> EConstr.constr

deprecated (9.2) Use Reductionops.nf_evar

val pr_gls : Proofview.Goal.t -> Pp.t

deprecated (9.2) Use Printer.pr_evar