Module UGraph

type t

Graphs of universes.

val set_type_in_type : bool -> t -> t

val type_in_type : t -> bool

When type_in_type, functions adding level constraints do not fail and may instead ignore inconsistent constraints.

Checking functions such as check_leq always return true.

type 'a check_function = t -> 'a -> 'a -> bool

val check_leq : Univ.Universe.t check_function

val check_eq : Univ.Universe.t check_function

val check_eq_level : Univ.Level.t check_function

val initial_universes : t

The initial graph of universes: Prop < Set

val initial_universes_with : t -> t

Initial universes, but keeping options such as type in type from the argument.

val check_eq_instances : (Sorts.Quality.t -> Sorts.Quality.t -> bool) -> t -> UVars.Instance.t -> UVars.Instance.t -> bool

Check equality of instances w.r.t. a universe graph

...

Merge of constraints in a universes graph. The function merge_constraints merges a set of constraints in a given universes graph. It raises the exception UniverseInconsistency if the constraints are not satisfiable.

type path_explanation

type explanation =

| Path of path_explanation

| Other of Pp.t

type univ_inconsistency = Sorts.printer option * (Univ.UnivConstraint.kind * Sorts.t * Sorts.t * explanation option)

exception UniverseInconsistency of univ_inconsistency

val enforce_constraint : Univ.UnivConstraint.t -> t -> t

val merge_constraints : Univ.UnivConstraints.t -> t -> t

val check_constraint : t -> Univ.UnivConstraint.t -> bool

val check_constraints : Univ.UnivConstraints.t -> t -> bool

val check_eq_sort : (Sorts.Quality.t -> Sorts.Quality.t -> bool) -> t -> Sorts.t -> Sorts.t -> bool

Checks whether (i) the first quality is equal to the second and (ii) that the universe of the first one is equal to the universe of the second one.

val check_leq_sort : (Sorts.Quality.t -> Sorts.Quality.t -> bool) -> t -> Sorts.t -> Sorts.t -> bool

Checks whether (i) the second quality eliminates into the first and (ii) that the universe of the first one is below the universe of the second one.

val enforce_leq_alg : Univ.Universe.t -> Univ.Universe.t -> t -> Univ.UnivConstraints.t * t

Adds a universe to the graph, ensuring it is >= or > Set.

raises AlreadyDeclared
if the level is already declared in the graph.

exception AlreadyDeclared

val add_universe : Univ.Level.t -> strict:bool -> t -> t

val check_declared_universes : t -> Univ.Level.Set.t -> (unit, Univ.Level.Set.t) Stdlib.result

Check that the universe levels are declared.

val empty_universes : t

The empty graph of universes

val constraints_of_universes : t -> Univ.UnivConstraints.t * Univ.Level.Set.t list

constraints_of_universes g returns csts and partition where csts are the non-Eq constraints and partition is the partition of the universes into equivalence classes.

val constraints_for : kept:Univ.Level.Set.t -> t -> Univ.UnivConstraints.t

constraints_for ~kept g returns the constraints about the universes kept in g up to transitivity.

eg if g is a <= b <= c then constraints_for ~kept:{a, c} g is a <= c.

val domain : t -> Univ.Level.Set.t

Known universes

Dumping

type node =

| Alias of Univ.Level.t

| Node of bool Univ.Level.Map.t
(*
Nodes v s.t. u < v (true) or u <= v (false)
*)

val repr : t -> node Univ.Level.Map.t

Pretty-printing of universes.

val pr_universes : (Univ.Level.t -> Pp.t) -> node Univ.Level.Map.t -> Pp.t

val explain_universe_inconsistency : Sorts.printer -> univ_inconsistency -> Pp.t

val check_universes_invariants : t -> unit

Debugging

module Internal : sig ... end