Module UVars.AbstractContext

type t

An abstract context serves to quantify over a graph of universes represented using de Bruijn indices, as in: u0, ..., u(n-1), Var i < Var j, .., Var k <= Var l |- term(Var 0 .. Var (n-1)) \-------------/ \-------------------------------/ \---------------------/ names for constraints expressed on de Bruijn judgement in abstract printing representation of the n univ vars context expected to use de Bruijn indices

val make : bound_names -> PConstraints.t -> t

Build an abstract context. PConstraints may be between sorts/universe variables.

val repr : t -> UContext.t

repr ctx is (Var(0), ... Var(n-1) |= cstr where n is the length of the context and cstr the abstracted UnivConstraints.t.

val hcons : t -> int * t

val empty : t

val is_empty : t -> bool

val is_constant : t -> bool

Empty instance, but may have constraints

val size : t -> int * int

val union : t -> t -> t

The constraints are expected to be relative to the concatenated set of universes

val instantiate : Instance.t -> t -> PConstraints.t

Generate the set of instantiated PConstraints.t *

val names : t -> bound_names

Return the names of the bound universe variables

val constraints : t -> PConstraints.t

Return the constraints on the universe variables

val refine_names : bound_names -> t -> t

Use names to name the possibly yet unnamed universes

val pr : Sorts.printer -> ?variance:Variance.t array -> t -> Pp.t