Module UVars.Variance

type t =

| Irrelevant

| Covariant

| Invariant

A universe position in the instance given to a cumulative inductive can be the following. Note there is no Contravariant case because forall x : A, B <= forall x : A', B' requires A = A' as opposed to A' <= A.

val check_subtype : t -> t -> bool

check_subtype x y holds if variance y is also an instance of x

val sup : t -> t -> t

val pr : t -> Pp.t

val equal : t -> t -> bool