Module Univ.ContextSet

type t = Level.Set.t * UnivConstraints.t

val empty : t

val is_empty : t -> bool

val singleton : Level.t -> t

val of_set : Level.Set.t -> t

val equal : t -> t -> bool

val union : t -> t -> t

val append : t -> t -> t

Variant of union which is more efficient when the left argument is much smaller than the right one.

val diff : t -> t -> t

val add_universe : Level.t -> t -> t

val add_constraints : UnivConstraints.t -> t -> t

val constraints : t -> UnivConstraints.t

val levels : t -> Level.Set.t

val size : t -> int

The number of universes in the context

val hcons : t Hashcons.f

val pr : (Level.t -> Pp.t) -> t -> Pp.t