Module Univ.Level

Universes.

type t

Type of universe levels. A universe level is essentially a unique name that will be associated to constraints later on. A level can be local to a definition or global.

val set : t

The Set universe level.

val is_set : t -> bool

Is the universe Set?

val compare : t -> t -> int

Comparison function

val equal : t -> t -> bool

Equality function

val hash : t -> int

val hcons : t Hashcons.f

val make : UGlobal.t -> t

val raw_pr : t -> Pp.t

Pretty-printing. When possible you should use name-aware printing.

val pr : t -> Pp.t

val to_string : t -> string

Debug printing

val var : int -> t

val var_index : t -> int option

val name : t -> UGlobal.t option

module Set : sig ... end

module Map : sig ... end