Module UnivMinim

module UPairSet : CSet.S with type elt = Univ.Level.t * Univ.Level.t

Unordered pairs of universe levels (ie (u,v) = (v,u))

type extra = {

weak_constraints : UPairSet.t;

above_prop : Univ.Level.Set.t;

}

val empty_extra : extra

val extra_union : extra -> extra -> extra

Simplification and pruning of constraints: normalize_context_set ctx us

Instantiate the variables in us with their most precise universe levels respecting the constraints.

Normalizes the context ctx w.r.t. equality constraints, choosing a canonical universe in each equivalence class (a global one if there is one) and transitively saturate the constraints w.r.t to the equalities.

val normalize_context_set : UGraph.t -> Univ.ContextSet.t -> UnivFlex.t -> extra -> UnivFlex.t * Univ.ContextSet.t