Module Util.Compare

Helpers to write comparison functions

type list =

| []

| :: : (('a -> 'a -> int) * 'a * 'a) * list -> list

val compare : list -> int

Compare the elements in the order they are given using the provided comparison function until the first non-zero result.