Module WriteOnceArray

type 'a t

Like a Parray, but the old pointers are invalidated instead of updated

val make : int -> 'a t

val get : int -> 'a t -> 'a option

val is_filled : int -> 'a t -> bool

val add : int -> 'a -> 'a t -> 'a t

val fill_remaining : 'a -> 'a t -> 'a t

val to_array : 'a t -> 'a array

val to_array_opt : 'a t -> 'a array option

The NoDupArray is still invalidated if the result is None

module Internal : sig ... end