Module AllScheme.Warning_scheme_all

type cache

type t

val empty_cache : unit -> cache

val warn : t -> cache -> unit

Warning for looking up the all predicate and its theorem. If this warning is alredy in the cache do nothing, oterwise warn and add it to the cache