Module Elimschemes

val declare_prop_but_default_dependent_elim : Names.inductive -> unit

Declare an inductive should be eliminated dependently even though it's in Prop

val is_prop_but_default_dependent_elim : Names.inductive -> bool

Check if an inductive should be eliminated dependently even though it's in Prop

val pseudo_sort_quality_for_elim : Names.inductive -> Declarations.one_inductive_body -> Sorts.Quality.t

Legacy API, returns the quality of the inductive except if it's prop_but_default_dependent_elim in which case we return Type instead.

val default_case_analysis_dependence : Environ.env -> Names.inductive -> bool

Returns false on inductives which cannot be eliminated dependently or are in Prop without being declared prop_but_default_dependent_elim.

val elim_scheme : dep:bool -> to_kind:UnivGen.QualityOrSet.t -> Ind_tables.individual Ind_tables.scheme_kind

Induction/recursion schemes

Case analysis schemes

val case_dep : Ind_tables.individual Ind_tables.scheme_kind

val case_nodep : Ind_tables.individual Ind_tables.scheme_kind

val casep_dep : Ind_tables.individual Ind_tables.scheme_kind

val casep_nodep : Ind_tables.individual Ind_tables.scheme_kind

val scase_dep : Ind_tables.individual Ind_tables.scheme_kind

val scase_nodep : Ind_tables.individual Ind_tables.scheme_kind

Recursor names utilities

val lookup_eliminator : Environ.env -> Names.inductive -> UnivGen.QualityOrSet.t -> Names.GlobRef.t

val elimination_suffix : UnivGen.QualityOrSet.t -> string

val make_elimination_ident : Names.Id.t -> UnivGen.QualityOrSet.t -> Names.Id.t