Module Esubst.Internal

Debugging utilities

type 'a or_rel =

| REL of int

| VAL of int * 'a

val repr : 'a subs -> 'a or_rel list * int

High-level representation of a substitution. The first component is a list that associates a value to an index, and the second component is the relocation shift that must be applied to any variable pointing outside of the substitution.