Module PrintingFlags.Extern

module FactorizeEqns : sig ... end

module Records : sig ... end

type t = {

use_implicit_types : bool;

records : Records.t;

implicits : bool;	(*
When implicits is on then implicits_explicit_args tells how implicit args are printed. If on, implicit args are printed with the form (id:=arg) otherwise arguments are printed normally and the function is prefixed by "@".
*)

implicits_explicit_args : bool;

implicits_defensive : bool;

coercions : bool;

parentheses : bool;

notations : bool;

raw_literals : bool;

projections : bool;

float : bool;

factorize_eqns : FactorizeEqns.t;

depth : int option;

}

val make_raw : t -> t

val current : unit -> t

val current_ignore_raw : unit -> t