Module Rocqshim

type worker = {

package : string;

basename : string;

}

val get_worker_path : worker -> string

Find the executable for the given worker. init must have been called. byte defaults to whether the current executable is byte compiled.

type opts = {

debug_shim : bool;

}

val parse_opts : string list -> opts * string list

val init : opts -> string list -> unit

Initialize environment and search paths.

val exec_or_create_process : string -> string array -> 'a

On windows Unix.execv creates a new process and exits this one. This confuses dune into thinking we are done, so instead we create_process and wait for it.