Module UnivGen

module QualityOrSet : sig ... end

type univ_length_mismatch = {

gref : Names.GlobRef.t;

actual : int * int;

expect : int * int;

}

exception UniverseLengthMismatch of univ_length_mismatch

Side-effecting functions creating new universe levels.

val new_univ_global : unit -> Univ.UGlobal.t

val new_sort_global : Names.Id.t -> Sorts.QGlobal.t

val fresh_level : unit -> Univ.Level.t

val fresh_sort_quality : unit -> Sorts.QVar.t

Build a fresh instance for a given context, its associated substitution and the instantiated constraints.

type sort_context_set = (Sorts.QVar.Set.t * Univ.Level.Set.t) * PConstraints.t

type 'a in_sort_context_set = 'a * sort_context_set

val sort_context_union : sort_context_set -> sort_context_set -> sort_context_set

val empty_sort_context : sort_context_set

val is_empty_sort_context : sort_context_set -> bool

val diff_sort_context : sort_context_set -> sort_context_set -> sort_context_set

val pr_sort_context : Sorts.printer -> sort_context_set -> Pp.t

val fresh_instance : UVars.AbstractContext.t -> UVars.Instance.t in_sort_context_set

val fresh_instance_from : ?loc:Loc.t -> UVars.AbstractContext.t -> (Names.GlobRef.t * UVars.Instance.t) option -> UVars.Instance.t in_sort_context_set

The globref is only used for the error message when there is a mismatch.

val fresh_sort_in_quality : QualityOrSet.t -> Sorts.t in_sort_context_set

NB: QSort is treated as QType

val fresh_constant_instance : Environ.env -> Names.Constant.t -> Constr.pconstant in_sort_context_set

val fresh_inductive_instance : Environ.env -> Names.inductive -> Constr.pinductive in_sort_context_set

val fresh_constructor_instance : Environ.env -> Names.constructor -> Constr.pconstructor in_sort_context_set

val fresh_array_instance : Environ.env -> UVars.Instance.t in_sort_context_set

val fresh_global_instance : ?loc:Loc.t -> ?names:UVars.Instance.t -> Environ.env -> Names.GlobRef.t -> Constr.constr in_sort_context_set

val fresh_sort_context_instance : sort_context_set -> UVars.sort_level_subst * sort_context_set

val constr_of_monomorphic_global : Environ.env -> Names.GlobRef.t -> Constr.constr

Create a fresh global in the environment argument, without side effects. BEWARE: this raises an error on polymorphic constants/inductives: the constraints should be properly added to an evd. See Evd.fresh_global, Evarutil.new_global, and pf_constr_of_global for the proper way to get a fresh copy of a polymorphic global reference.